==> scripts/0.1 Definición directorios.R <==

# Definimos directorios de trabajo
directorio_in_EUSTAT<- "C:/Users/Leire/OneDrive/Escritorio/EUSTAT-Turismo/Datos/0 input EUSTAT"
directorio_in_OTROS<- "C:/Users/Leire/OneDrive/Escritorio/EUSTAT-Turismo/Datos/1 input OTROS"
directorio_out<-"C:/Users/Leire/OneDrive/Escritorio/EUSTAT-Turismo/Datos/output"
