==> scripts/0.2 CÓDIGO BASE.R <==

# 0.0 Codigo Base
##################################################################################

# Instalación de librerías
source("C:/Users/Leire/OneDrive/Escritorio/EUSTAT-Turismo/scripts/0.0 Carga librerías.R")

# Definición de directorios
source("C:/Users/Leire/OneDrive/Escritorio/EUSTAT-Turismo/scripts/0.1 Definición Directorios.R")

# Carga datos ENCUESTA
source("C:/Users/Leire/OneDrive/Escritorio/EUSTAT-Turismo/scripts/1 Carga datos ENCUESTA.R")

# Carga datos BOOKING
source("C:/Users/Leire/OneDrive/Escritorio/EUSTAT-Turismo/scripts/2 Carga datos BOOKING.R")
